Resolvendo a autenticação

+++ server/database/migrations/2023_06_02_122340_create_expense_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('merchandise_purchased')->nullable(false);
             $table->string('establishment')->nullable(false);
             $table->integer('installments')->nullable(false);
-            $table->double('value_installment')->nullable(false);
+            $table->float('value_installment',10,2)->nullable(false);
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
             $table->timestamps();
         });
